+++ cmd/spool-rack/main.go
@@ -21,7 +21,7 @@ const (
 )
 
 var (
-	version = "0.3.0"
+	version = "0.4.0"
 )
 
 func main() {